+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".poc.PocActivity">
+            android:name=".tasks.PocActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".addedittask.AddEditTaskActivity">
+        <activity android:name=".addedittask.AddTaskActivity">
 
         </activity>
