--- app/src/main/java/com/ntg/user/mvpsample/BaseView.java
@@ -1,9 +1,0 @@
-package com.ntg.user.mvpsample;
-
-/**
- * Created by mohamed yassin on 1/27/2018.
- */
-
-public interface BaseView<T> {
-    void setPresenter(T presenter);
-}
