--- app/src/main/java/com/ntg/user/mvpsample/StatisticsActivity.java
@@ -1,8 +1,0 @@
-package com.ntg.user.mvpsample;
-
-/**
- * Created by mohamed yassin on 1/29/2018.
- */
-
-public class StatisticsActivity {
-}
