--- app/src/main/java/com/ntg/user/mvpsample/addedittask/AddEditTaskActivity.java
@@ -1,87 +1,0 @@
-package com.ntg.user.mvpsample.addedittask;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
-import com.ntg.user.mvpsample.ActivityUtils;
-import com.ntg.user.mvpsample.Injection;
-import com.ntg.user.mvpsample.R;
-
-/**
- * Created by mohamed yassin on 1/28/2018.
- */
-
-public class AddEditTaskActivity extends AppCompatActivity {
-    public static final int REQUEST_ADD_TASK = 1;
-
-    public static final String SHOULD_LOAD_DATA_FROM_REPO_KEY = "SHOULD_LOAD_DATA_FROM_REPO_KEY";
-
-    private AddEditTaskPresenter mAddEditTaskPresenter;
-
-    private ActionBar mActionBar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.addtask_act);
-
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        mActionBar = getSupportActionBar();
-        mActionBar.setDisplayHomeAsUpEnabled(true);
-        mActionBar.setDisplayShowHomeEnabled(true);
-
-        AddEditTaskFragment addEditTaskFragment = (AddEditTaskFragment) getSupportFragmentManager()
-                .findFragmentById(R.id.contentFrame);
-
-        String taskId = getIntent().getStringExtra(AddEditTaskFragment.ARGUMENT_EDIT_TASK_ID);
-
-        setToolbarTitle(taskId);
-
-        if (addEditTaskFragment == null) {
-            addEditTaskFragment = AddEditTaskFragment.newInstance();
-
-            if (getIntent().hasExtra(AddEditTaskFragment.ARGUMENT_EDIT_TASK_ID)) {
-                Bundle bundle = new Bundle();
-                bundle.putString(AddEditTaskFragment.ARGUMENT_EDIT_TASK_ID, taskId);
-                addEditTaskFragment.setArguments(bundle);
-            }
-
-            ActivityUtils.addFragmentToActivity(getSupportFragmentManager(),
-                    addEditTaskFragment, R.id.contentFrame);
-        }
-
-        boolean shouldLoadDataFromRepo = true;
-
-        // Prevent the presenter from loading data from the repository if this is a config change.
-        if (savedInstanceState != null) {
-            // Data might not have loaded when the config change happen, so we saved the state.
-            shouldLoadDataFromRepo = savedInstanceState.getBoolean(SHOULD_LOAD_DATA_FROM_REPO_KEY);
-        }
-
-        // Create the presenter
-        mAddEditTaskPresenter = new AddEditTaskPresenter(
-                taskId,
-                Injection.provideTasksRepository(getApplicationContext()),
-                addEditTaskFragment,
-                shouldLoadDataFromRepo);
-    }
-    private void setToolbarTitle(@Nullable String taskId) {
-        if (taskId == null) {
-            mActionBar.setTitle(R.string.add_task);
-        } else {
-            mActionBar.setTitle(R.string.edit_task);
-        }
-    }
-
-    @Override
-    protected void onSaveInstanceState(Bundle outState) {
-        // Save the state so that next time we know if we need to refresh data.
-        outState.putBoolean(SHOULD_LOAD_DATA_FROM_REPO_KEY, mAddEditTaskPresenter.isDataMissing());
-        super.onSaveInstanceState(outState);
-    }
-}
