+++ app/src/main/java/com/ntg/user/mvpsample/addedittask/AddEditTaskContract.java
@@ -1,32 +1,20 @@
 package com.ntg.user.mvpsample.addedittask;
 
 import com.ntg.user.mvpsample.BasePresenter;
-import com.ntg.user.mvpsample.BaseView;
 
 /**
  * Created by mohamed yassin on 1/28/2018.
  */
 
-public interface AddEditTaskContract  {
+public interface AddEditTaskContract {
     interface View extends BaseView<Presenter> {
 
         void showEmptyTaskError();
 
         void showTasksList();
-
-        void setTitle(String title);
-
-        void setDescription(String description);
-
-        boolean isActive();
     }
+
     interface Presenter extends BasePresenter {
-
         void saveTask(String title, String description);
-
-        void populateTask();
-
-        boolean isDataMissing();
     }
-
 }
