+++ app/src/main/java/com/ntg/user/mvpsample/addedittask/AddEditTaskPresenter.java
@@ -1,11 +1,9 @@
 package com.ntg.user.mvpsample.addedittask;
 
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 
-import com.ntg.user.mvpsample.data.Task;
-import com.ntg.user.mvpsample.data.source.TasksDataSource;
-import com.ntg.user.mvpsample.data.source.TasksRepository;
+import com.ntg.user.mvpsample.Task;
+import com.ntg.user.mvpsample.addedittask.data.source.TasksDataSource;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -14,99 +12,36 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * Created by mohamed yassin on 1/28/2018.
  */
 
-public class AddEditTaskPresenter implements AddEditTaskContract.Presenter,
-        TasksDataSource.GetTaskCallback {
+public class AddEditTaskPresenter implements AddEditTaskContract.Presenter {
     @NonNull
     private final TasksDataSource mTasksRepository;
-
     @NonNull
     private final AddEditTaskContract.View mAddTaskView;
 
-    @Nullable
-    private String mTaskId;
 
-    private boolean mIsDataMissing;
+    // Creates a presenter for the add/edit view.
 
+    public AddEditTaskPresenter(@NonNull TasksDataSource tasksRepository,
+                                @NonNull AddEditTaskContract.View addTaskView) {
 
-     // Creates a presenter for the add/edit view.
-
-    public AddEditTaskPresenter(@Nullable String taskId, @NonNull TasksDataSource tasksRepository,
-                                @NonNull AddEditTaskContract.View addTaskView, boolean shouldLoadDataFromRepo) {
-        mTaskId = taskId;
         mTasksRepository = checkNotNull(tasksRepository);
         mAddTaskView = checkNotNull(addTaskView);
-        mIsDataMissing = shouldLoadDataFromRepo;
-
         mAddTaskView.setPresenter(this);
     }
 
     @Override
     public void start() {
-        if (!isNewTask() && mIsDataMissing) {
-            populateTask();
-        }
     }
 
     @Override
     public void saveTask(String title, String description) {
-        if (isNewTask()) {
-            createTask(title, description);
-        } else {
-            updateTask(title, description);
-        }
-    }
-
-    @Override
-    public void populateTask() {
-        if (isNewTask()) {
-            throw new RuntimeException("populateTask() was called but task is new.");
-        }
-        mTasksRepository.getTask(mTaskId, this);
-    }
-
-    @Override
-    public void onTaskLoaded(Task task) {
-        // The view may not be able to handle UI updates anymore
-        if (mAddTaskView.isActive()) {
-            mAddTaskView.setTitle(task.getTitle());
-            mAddTaskView.setDescription(task.getDescription());
-        }
-        mIsDataMissing = false;
-    }
-
-    @Override
-    public void onDataNotAvailable() {
-        // The view may not be able to handle UI updates anymore
-        if (mAddTaskView.isActive()) {
-            mAddTaskView.showEmptyTaskError();
-        }
-    }
-
-    @Override
-    public boolean isDataMissing() {
-        return mIsDataMissing;
-    }
-
-    private boolean isNewTask() {
-        return mTaskId == null;
-    }
-
-    private void createTask(String title, String description) {
         Task newTask = new Task(title, description);
-        if (newTask.isEmpty()) {
+        if (newTask.getTitle().isEmpty()) {
             mAddTaskView.showEmptyTaskError();
         } else {
             mTasksRepository.saveTask(newTask);
             mAddTaskView.showTasksList();
         }
-    }
 
-    private void updateTask(String title, String description) {
-        if (isNewTask()) {
-            throw new RuntimeException("updateTask() was called but task is new.");
-        }
-        mTasksRepository.saveTask(new Task(title, description));
-        mAddTaskView.showTasksList(); // After an edit, go back to the list.
     }
-
 }
