+++ app/src/main/java/com/ntg/user/mvpsample/addedittask/data/source/TasksRepository.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2016, The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.ntg.user.mvpsample.addedittask.data.source;
+
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+
+
+import com.ntg.user.mvpsample.Task;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+
+public class TasksRepository implements TasksDataSource {
+
+    private static TasksRepository INSTANCE = null;
+
+    private final TasksDataSource mTasksRemoteDataSource;
+
+    private final TasksDataSource mTasksLocalDataSource;
+
+    private TasksRepository(@NonNull TasksDataSource tasksRemoteDataSource,
+                            @NonNull TasksDataSource tasksLocalDataSource) {
+        mTasksRemoteDataSource = checkNotNull(tasksRemoteDataSource);
+        mTasksLocalDataSource = checkNotNull(tasksLocalDataSource);
+    }
+
+
+    public static TasksRepository getInstance(TasksDataSource tasksRemoteDataSource,
+                                              TasksDataSource tasksLocalDataSource) {
+        if (INSTANCE == null) {
+            INSTANCE = new TasksRepository(tasksRemoteDataSource, tasksLocalDataSource);
+        }
+        return INSTANCE;
+    }
+
+    @Override
+    public void saveTask(@NonNull Task task) {
+        mTasksRemoteDataSource.saveTask(task);
+        mTasksLocalDataSource.saveTask(task);
+    }
+}
+
