+++ app/src/main/java/com/ntg/user/mvpsample/addedittask/data/source/local/TasksLocalDataSource.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2016, The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.ntg.user.mvpsample.addedittask.data.source.local;
+
+import android.content.ContentValues;
+import android.content.Context;
+import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
+import android.support.annotation.NonNull;
+
+
+import com.ntg.user.mvpsample.Task;
+import com.ntg.user.mvpsample.addedittask.data.source.TasksDataSource;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+
+/**
+ * Concrete implementation of a data source as a db.
+ */
+public class TasksLocalDataSource implements TasksDataSource {
+
+    private static TasksLocalDataSource INSTANCE;
+
+    private TasksDbHelper mDbHelper;
+
+    // Prevent direct instantiation.
+    private TasksLocalDataSource(@NonNull Context context) {
+        checkNotNull(context);
+        mDbHelper = new TasksDbHelper(context);
+    }
+
+    public static TasksLocalDataSource getInstance(@NonNull Context context) {
+        if (INSTANCE == null) {
+            INSTANCE = new TasksLocalDataSource(context);
+        }
+        return INSTANCE;
+    }
+
+    @Override
+    public void saveTask(@NonNull Task task) {
+        checkNotNull(task);
+        SQLiteDatabase db = mDbHelper.getWritableDatabase();
+        ContentValues values = new ContentValues();
+        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID, task.getUserId());
+        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE, task.getTitle());
+        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION, task.getDesc());
+        db.insert(TasksPersistenceContract.TaskEntry.TABLE_NAME, null, values);
+        db.close();
+    }
+}
