--- app/src/main/java/com/ntg/user/mvpsample/data/source/TasksDataSource.java
@@ -1,49 +1,0 @@
-package com.ntg.user.mvpsample.data.source;
-
-import android.support.annotation.NonNull;
-
-import com.ntg.user.mvpsample.data.Task;
-
-import java.util.List;
-
-/**
- * Created by mohamed yassin on 1/28/2018.
- */
-
-public interface TasksDataSource {
-    interface LoadTasksCallback {
-
-        void onTasksLoaded(List<Task> tasks);
-
-        void onDataNotAvailable();
-    }
-
-    interface GetTaskCallback {
-
-        void onTaskLoaded(Task task);
-
-        void onDataNotAvailable();
-    }
-
-    void getTasks(@NonNull LoadTasksCallback callback);
-
-    void getTask(@NonNull String taskId, @NonNull GetTaskCallback callback);
-
-    void saveTask(@NonNull Task task);
-
-    void completeTask(@NonNull Task task);
-
-    void completeTask(@NonNull String taskId);
-
-    void activateTask(@NonNull Task task);
-
-    void activateTask(@NonNull String taskId);
-
-    void clearCompletedTasks();
-
-    void refreshTasks();
-
-    void deleteAllTasks();
-
-    void deleteTask(@NonNull String taskId);
-}
