--- app/src/main/java/com/ntg/user/mvpsample/data/source/local/TasksLocalDataSource.java
@@ -1,245 +1,0 @@
-/*
- * Copyright 2016, The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ntg.user.mvpsample.data.source.local;
-
-import android.content.ContentValues;
-import android.content.Context;
-import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.support.annotation.NonNull;
-
-
-import com.ntg.user.mvpsample.data.Task;
-import com.ntg.user.mvpsample.data.source.TasksDataSource;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-
-/**
- * Concrete implementation of a data source as a db.
- */
-public class TasksLocalDataSource implements TasksDataSource {
-
-    private static TasksLocalDataSource INSTANCE;
-
-    private TasksDbHelper mDbHelper;
-
-    // Prevent direct instantiation.
-    private TasksLocalDataSource(@NonNull Context context) {
-        checkNotNull(context);
-        mDbHelper = new TasksDbHelper(context);
-    }
-
-    public static TasksLocalDataSource getInstance(@NonNull Context context) {
-        if (INSTANCE == null) {
-            INSTANCE = new TasksLocalDataSource(context);
-        }
-        return INSTANCE;
-    }
-
-    /**
-     * Note: {@link LoadTasksCallback#onDataNotAvailable()} is fired if the database doesn't exist
-     * or the table is empty.
-     */
-    @Override
-    public void getTasks(@NonNull LoadTasksCallback callback) {
-        List<Task> tasks = new ArrayList<Task>();
-        SQLiteDatabase db = mDbHelper.getReadableDatabase();
-
-        String[] projection = {
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED
-        };
-
-        Cursor c = db.query(
-                TasksPersistenceContract.TaskEntry.TABLE_NAME, projection, null, null, null, null, null);
-
-        if (c != null && c.getCount() > 0) {
-            while (c.moveToNext()) {
-                String itemId = c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID));
-                String title = c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE));
-                String description =
-                        c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION));
-                boolean completed =
-                        c.getInt(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED)) == 1;
-              //  Task task = new Task(title, description, itemId, completed);
-                //tasks.add(task);
-            }
-        }
-        if (c != null) {
-            c.close();
-        }
-
-        db.close();
-
-        if (tasks.isEmpty()) {
-            // This will be called if the table is new or just empty.
-            callback.onDataNotAvailable();
-        } else {
-            callback.onTasksLoaded(tasks);
-        }
-
-    }
-
-    /**
-     * Note: {@link GetTaskCallback#onDataNotAvailable()} is fired if the {@link Task} isn't
-     * found.
-     */
-    @Override
-    public void getTask(@NonNull String taskId, @NonNull GetTaskCallback callback) {
-        SQLiteDatabase db = mDbHelper.getReadableDatabase();
-
-        String[] projection = {
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION,
-                TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED
-        };
-
-        String selection = TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID + " LIKE ?";
-        String[] selectionArgs = { taskId };
-
-        Cursor c = db.query(
-                TasksPersistenceContract.TaskEntry.TABLE_NAME, projection, selection, selectionArgs, null, null, null);
-
-        Task task = null;
-
-        if (c != null && c.getCount() > 0) {
-            c.moveToFirst();
-            String itemId = c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID));
-            String title = c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE));
-            String description =
-                    c.getString(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION));
-            boolean completed =
-                    c.getInt(c.getColumnIndexOrThrow(TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED)) == 1;
-         //   task = new Task(title, description, itemId, completed);
-        }
-        if (c != null) {
-            c.close();
-        }
-
-        db.close();
-
-        if (task != null) {
-            callback.onTaskLoaded(task);
-        } else {
-            callback.onDataNotAvailable();
-        }
-    }
-
-    @Override
-    public void saveTask(@NonNull Task task) {
-        checkNotNull(task);
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID, task.getUserId());
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_TITLE, task.getTitle());
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_DESCRIPTION, task.getDescription());
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED, task.isCompleted());
-
-        db.insert(TasksPersistenceContract.TaskEntry.TABLE_NAME, null, values);
-
-        db.close();
-    }
-
-    @Override
-    public void completeTask(@NonNull Task task) {
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED, true);
-
-        String selection = TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID + " LIKE ?";
-        String[] selectionArgs = { task.getUserId() };
-
-       db.update(TasksPersistenceContract.TaskEntry.TABLE_NAME, values, selection, selectionArgs);
-
-        db.close();
-    }
-
-    @Override
-    public void completeTask(@NonNull String taskId) {
-        // Not required for the local data source because the {@link TasksRepository} handles
-        // converting from a {@code taskId} to a {@link task} using its cached data.
-    }
-
-    @Override
-    public void activateTask(@NonNull Task task) {
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        ContentValues values = new ContentValues();
-        values.put(TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED, false);
-
-        String selection = TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID + " LIKE ?";
-      //  String[] selectionArgs = { task.getId() };
-
-      //  db.update(TasksPersistenceContract.TaskEntry.TABLE_NAME, values, selection, selectionArgs);
-
-        db.close();
-    }
-
-    @Override
-    public void activateTask(@NonNull String taskId) {
-        // Not required for the local data source because the {@link TasksRepository} handles
-        // converting from a {@code taskId} to a {@link task} using its cached data.
-    }
-
-    @Override
-    public void clearCompletedTasks() {
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        String selection = TasksPersistenceContract.TaskEntry.COLUMN_NAME_COMPLETED + " LIKE ?";
-        String[] selectionArgs = { "1" };
-
-        db.delete(TasksPersistenceContract.TaskEntry.TABLE_NAME, selection, selectionArgs);
-
-        db.close();
-    }
-
-    @Override
-    public void refreshTasks() {
-        // Not required because the {@link TasksRepository} handles the logic of refreshing the
-        // tasks from all the available data sources.
-    }
-
-    @Override
-    public void deleteAllTasks() {
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        db.delete(TasksPersistenceContract.TaskEntry.TABLE_NAME, null, null);
-
-        db.close();
-    }
-
-    @Override
-    public void deleteTask(@NonNull String taskId) {
-        SQLiteDatabase db = mDbHelper.getWritableDatabase();
-
-        String selection = TasksPersistenceContract.TaskEntry.COLUMN_NAME_ENTRY_ID + " LIKE ?";
-        String[] selectionArgs = { taskId };
-
-        db.delete(TasksPersistenceContract.TaskEntry.TABLE_NAME, selection, selectionArgs);
-
-        db.close();
-    }
-}
