--- app/src/main/java/com/ntg/user/mvpsample/poc/PocItemListener.java
@@ -1,10 +1,0 @@
-package com.ntg.user.mvpsample.poc;
-
-
-import com.ntg.user.mvpsample.poc.data.Poc;
-
-public interface PocItemListener {
-
-    void onPocClicked(Poc poc);
-
-}
