--- app/src/main/java/com/ntg/user/mvpsample/poc/data/PocDataSource.java
@@ -1,16 +1,0 @@
-package com.ntg.user.mvpsample.poc.data;
-
-import java.util.List;
-
-/**
- * Created by mohamed yassin on 1/29/2018.
- */
-
-public interface PocDataSource {
-    interface GetPocsCallBack {
-        void onPocsLoaded(List<Poc> pocList);
-
-        void onPocsFailed(String errMsg);
-    }
-    void getPocs(GetPocsCallBack getPocsCallBack);
-}
