+++ app/src/main/java/com/ntg/user/mvpsample/tasks/PocItemListener.java
@@ -1,0 +1,10 @@
+package com.ntg.user.mvpsample.tasks;
+
+
+import com.ntg.user.mvpsample.tasks.data.Poc;
+
+public interface PocItemListener {
+
+    void onPocClicked(Poc poc);
+
+}
