+++ app/src/main/res/layout/addtask_act.xml
@@ -7,7 +7,7 @@
     android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".addedittask.AddEditTaskActivity">
+    tools:context=".addedittask.AddTaskActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
